Decomposes easy-peasy.js into multiple files
