+++ src/create-store.js
@@ -1,0 +1,158 @@
+import {
+  applyMiddleware,
+  compose as reduxCompose,
+  createStore as reduxCreateStore,
+} from 'redux'
+import reduxThunk from 'redux-thunk'
+import { get } from './lib'
+import { metaSymbol } from './constants'
+import * as helpers from './helpers'
+import createStoreInternals from './create-store-internals'
+
+export default function createStore(model, options = {}) {
+  const {
+    compose,
+    devTools = true,
+    disableInternalSelectFnMemoize = false,
+    initialState = {},
+    injections,
+    mockActions = false,
+    middleware = [],
+    reducerEnhancer = rootReducer => rootReducer,
+  } = options
+
+  const modelDefinition = {
+    ...model,
+    logFullState: helpers.thunk((actions, payload, { getState }) => {
+      // eslint-disable-next-line no-console
+      console.log(JSON.stringify(getState(), null, 2))
+    }),
+    replaceState: helpers.action((state, payload) => payload),
+  }
+
+  const references = {}
+
+  let mockedActions = []
+
+  const mockActionsMiddlware = () => next => action => {
+    if (mockActions) {
+      mockedActions.push(action)
+      return undefined
+    }
+    return next(action)
+  }
+
+  const dispatchThunkListeners = (name, payload) => {
+    const listensForAction = references.internals.thunkListenersDict[name]
+    return listensForAction && listensForAction.length > 0
+      ? Promise.all(
+          listensForAction.map(listenForAction =>
+            listenForAction(
+              get(
+                listenForAction[metaSymbol].parent,
+                references.internals.actionCreators,
+              ),
+              payload,
+              {
+                dispatch: references.dispatch,
+                getState: references.getState,
+                injections,
+                meta: listenForAction[metaSymbol],
+              },
+            ),
+          ),
+        )
+      : Promise.resolve()
+  }
+
+  const dispatchActionStringListeners = () => next => action => {
+    if (references.internals.thunkListenersDict[action.type]) {
+      dispatchThunkListeners(action.type, action.payload)
+    }
+    return next(action)
+  }
+
+  const composeEnhancers =
+    compose ||
+    (devTools &&
+    typeof window !== 'undefined' &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+      ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+      : reduxCompose)
+
+  const bindStoreInternals = state => {
+    references.internals = createStoreInternals({
+      disableInternalSelectFnMemoize,
+      initialState: state,
+      injections,
+      model: modelDefinition,
+      references,
+      reducerEnhancer,
+    })
+  }
+
+  bindStoreInternals(initialState)
+
+  const store = reduxCreateStore(
+    references.internals.reducer,
+    references.internals.defaultState,
+    composeEnhancers(
+      applyMiddleware(
+        reduxThunk,
+        dispatchActionStringListeners,
+        mockActionsMiddlware,
+        ...middleware,
+      ),
+    ),
+  )
+
+  store.getMockedActions = () => [...mockedActions]
+  store.clearMockedActions = () => {
+    mockedActions = []
+  }
+
+  references.dispatch = store.dispatch
+  references.getState = store.getState
+
+  // attach the action creators to dispatch
+  const bindActionCreators = actionCreators => {
+    Object.keys(store.dispatch).forEach(actionsKey => {
+      delete store.dispatch[actionsKey]
+    })
+    Object.keys(actionCreators).forEach(key => {
+      store.dispatch[key] = actionCreators[key]
+    })
+  }
+
+  bindActionCreators(references.internals.actionCreators)
+
+  const rebuildStore = () => {
+    bindStoreInternals(store.getState())
+    store.replaceReducer(references.internals.reducer)
+    store.dispatch.replaceState(references.internals.defaultState)
+    bindActionCreators(references.internals.actionCreators)
+    return store
+  }
+
+  store.addModel = (key, modelForKey) => {
+    if (modelDefinition[key]) {
+      throw new Error(
+        `The store model already contains a model definition for "${key}"`,
+      )
+    }
+    modelDefinition[key] = modelForKey
+    return rebuildStore()
+  }
+
+  store.removeModel = key => {
+    if (!modelDefinition[key]) {
+      throw new Error(
+        `The store model does not contains a model definition for "${key}"`,
+      )
+    }
+    delete modelDefinition[key]
+    return rebuildStore()
+  }
+
+  return store
+}
